Fixed codacy and changes on tests suite.
