--- tests/s3/multipart_upload/test_multipart_uploadpartcopy.py
@@ -1,760 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-#
-# Copyright (c) 2020 Seagate Technology LLC and/or its Affiliates
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# For any questions about this software or licensing,
-# please email opensource@seagate.com or cortx-questions@seagate.com.
-
-"""Multipart upload PartCopy object test module."""
-
-import logging
-import random
-import os
-from time import perf_counter_ns
-
-import pytest
-from commons.ct_fail_on import CTFailOn
-from commons.errorcodes import error_handler
-from commons.exceptions import CTException
-from commons.utils.system_utils import create_file, remove_file, path_exists
-from commons.utils.s3_utils import get_precalculated_parts
-from commons.utils.system_utils import make_dirs, remove_dirs
-from commons.utils import assert_utils
-from commons.utils import s3_utils
-from commons.params import TEST_DATA_FOLDER
-from libs.s3 import S3_CFG
-from config.s3 import MPART_CFG
-from libs.s3.s3_common_test_lib import S3BackgroundIO
-from libs.s3 import s3_common_test_lib as cmn_lib
-from libs.s3.s3_multipart_test_lib import S3MultipartTestLib
-from libs.s3.s3_test_lib import S3TestLib
-from libs.s3.s3_restapi_test_lib import S3AccountOperationsRestAPI
-
-
-class TestMultipartUploadPartCopy:
-    """Multipart Upload PartCopy Test Suite."""
-
-    @classmethod
-    def setup_class(cls):
-        """
-        This is called only once before starting any tests in this class
-        """
-        cls.log = logging.getLogger(__name__)
-        cls.s3_test_obj = S3TestLib()
-        cls.s3_mpu_test_obj = S3MultipartTestLib(endpoint_url=S3_CFG["s3_url"])
-        cls.test_file = "mpu_obj"
-        cls.test_file_partcopy = "mpu_partcopy_obj"
-        cls.test_dir_path = os.path.join(TEST_DATA_FOLDER, "TestMultipartUploadRedesign")
-        cls.mp_obj_path = os.path.join(cls.test_dir_path, cls.test_file)
-        cls.mp_obj_path_partcopy = os.path.join(cls.test_dir_path, cls.test_file_partcopy)
-        cls.config_backup_path = os.path.join(cls.test_dir_path, "config_backup")
-        if not path_exists(cls.test_dir_path):
-            make_dirs(cls.test_dir_path)
-            cls.log.info("Created path: %s", cls.test_dir_path)
-        cls.downloaded_file = "{}{}".format("get_blackboxs3obj", perf_counter_ns())
-        cls.downloaded_file_path = os.path.join(
-            cls.test_dir_path, cls.downloaded_file)
-
-    @classmethod
-    def teardown_class(cls):
-        """
-        This is called after all tests in this class finished execution
-        """
-        cls.log.info("STARTED: teardown test suite operations.")
-        if path_exists(cls.test_dir_path):
-            remove_dirs(cls.test_dir_path)
-        cls.log.info("Cleanup test directory: %s", cls.test_dir_path)
-        cls.log.info("ENDED: teardown test suite operations.")
-
-    def setup_method(self):
-        """This is called before each test in this test suite."""
-        self.log.info("STARTED: Setup operations")
-        self.acc_name_prefix = "mpu-s3acc-{}"
-        self.acc_email_prefix = "{}@seagate.com"
-        self.s3acc_passwd = S3_CFG["CliConfig"]["s3_account"]["password"]
-        self.bucket_name = "mpu-bkt-{}".format(perf_counter_ns())
-        self.io_bucket_name = "mpu-io-bkt-{}".format(perf_counter_ns())
-        self.bucket_name1 = "mpu-bkt1-{}".format(perf_counter_ns())
-        self.bucket_name2 = "mpu-bkt2-{}".format(perf_counter_ns())
-        self.object_name = "mpu-obj-{}".format(perf_counter_ns())
-        self.object_name1 = "mpu-obj1-{}".format(perf_counter_ns())
-        self.object_name2 = "mpu-obj2-{}".format(perf_counter_ns())
-        self.mpu_partcopy_bkt = "mpu-partcopy-bkt-{}".format(perf_counter_ns())
-        self.mpu_partcopy_obj = "mpu-partcopy-obj-{}".format(perf_counter_ns())
-        self.s3bio_obj = S3BackgroundIO(self.s3_test_obj, self.io_bucket_name)
-        self.s3mpu_obj = S3MultipartTestLib()
-        self.s3t_obj = S3TestLib()
-        self.rest_obj = S3AccountOperationsRestAPI()
-        self.account_list = []
-        # create bucket
-        self.log.info("Creating a bucket with name : %s", self.bucket_name)
-        resp = self.s3_test_obj.create_bucket(self.bucket_name)
-        assert_utils.assert_true(resp[0], resp[1])
-        assert_utils.assert_equal(resp[1], self.bucket_name, resp[1])
-        self.log.info("Created a bucket with name : %s", self.bucket_name)
-        self.log.info("ENDED: Setup operations")
-
-    def teardown_method(self):
-        """This is called after each test in this test suite."""
-        self.log.info("STARTED: Teardown operations")
-        self.s3bio_obj.cleanup()
-        resp = self.s3_test_obj.bucket_list()
-        pref_list = [
-            each_bucket for each_bucket in resp[1] if each_bucket.startswith("mpu-bkt")]
-        if pref_list:
-            resp = self.s3_test_obj.delete_multiple_buckets(pref_list)
-            assert_utils.assert_true(resp[0], resp[1])
-        if path_exists(self.mp_obj_path):
-            remove_file(self.mp_obj_path)
-        for s3account in self.account_list:
-            resp = self.rest_obj.delete_s3_account(s3account)
-            assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Deleted a backup file and directory")
-        self.log.info("ENDED: Teardown operations")
-
-    def create_and_complete_mpu(self, mpu_cfg):
-        mpu_id = self.initiate_multipart(self.bucket_name, self.object_name)
-        uploaded_parts = get_precalculated_parts(self.mp_obj_path, mpu_cfg["part_sizes"],
-                                                 chunk_size=mpu_cfg["chunk_size"])
-        keys = list(uploaded_parts.keys())
-        random.shuffle(keys)
-        self.log.info("Uploading parts")
-        status, new_parts = self.s3_mpu_test_obj.upload_parts_sequential(mpu_id, self.bucket_name,
-                                                                         self.object_name,
-                                                                         parts=uploaded_parts)
-        assert_utils.assert_true(status, f"Failed to upload parts: {new_parts}")
-        self.log.info("Listing parts of multipart upload")
-        res = self.s3_mpu_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res[1])
-        self.log.info("Listed parts of multipart upload: %s", res[1])
-        sorted_part_list = sorted(new_parts, key=lambda x: x['PartNumber'])
-        self.log.info("Complete the multipart ")
-        try:
-            resp = self.s3_mpu_test_obj.complete_multipart_upload(mpu_id, sorted_part_list,
-                                                                  self.bucket_name,
-                                                                  self.object_name)
-            assert_utils.assert_false(resp[0], resp[1])
-        except CTException as error:
-            self.log.error(error)
-            self.log.info("Failed to complete the multipart with provided part details")
-        upload_etag = resp[1]["ETag"]
-        self.log.info("Get the uploaded object")
-        status, res = self.s3_test_obj.get_object(self.bucket_name, self.object_name)
-        assert_utils.assert_true(status, res)
-        get_etag = res['ETag']
-        self.log.info("Compare ETags")
-        assert_utils.assert_equal(resp[1]["ETag"], get_etag,
-                                  f"Failed to match ETag: {upload_etag}, {get_etag}")
-        self.log.info("Matched ETag: %s, %s", upload_etag, get_etag)
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32702')
-    @CTFailOn(error_handler)
-    def test_mpu_upload_partcopy_32702(self):
-        """
-        This test is for getting object uploaded using multipart UploadPartCopy api
-        """
-        mp_config = MPART_CFG["test_32702_1"]
-        multipart_obj_path = self.mp_obj_path
-        if os.path.exists(multipart_obj_path):
-            os.remove(multipart_obj_path)
-        create_file(multipart_obj_path, mp_config["file_size"])
-        mp_config_2 = MPART_CFG["test_32702_2"]
-        if os.path.exists(self.mp_obj_path_partcopy):
-            os.remove(self.mp_obj_path_partcopy)
-        create_file(self.mp_obj_path_partcopy, mp_config_2["file_size"])
-        self.log.info("STARTED: Test get the object uploaded using MPU UploadPartCopy")
-        s3_background_io = S3BackgroundIO(s3_test_lib_obj=self.s3_test_obj)
-        self.log.info("start s3 IO's")
-        s3_background_io.start(log_prefix="TEST-32702_s3bench_ios", duration="0h2m")
-        # create and complete MPU for Bucket1/object1
-        self.create_and_complete_mpu(mp_config)
-        self.log.info("Creating a bucket with name : %s", self.mpu_partcopy_bkt)
-        res = self.s3_test_obj.create_bucket(self.mpu_partcopy_bkt)
-        assert_utils.assert_true(res[0], res[1])
-        assert_utils.assert_equal(res[1], self.mpu_partcopy_bkt, res[1])
-        self.log.info("Created a bucket with name : %s", self.mpu_partcopy_bkt)
-        mpu_id2 = self.initiate_multipart(self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        uploaded_parts2 = get_precalculated_parts(self.mp_obj_path_partcopy,
-                                                  mp_config_2["part_sizes"],
-                                                  chunk_size=mp_config_2["chunk_size"])
-        keys = list(uploaded_parts2.keys())
-        uploaded_parts2['3'] = uploaded_parts2.pop('2') # modified the part 2 as part 3
-        random.shuffle(keys)
-        self.log.info("Uploading parts")
-        status, new_parts = self.s3_mpu_test_obj.upload_parts_sequential(mpu_id2,
-                                                                         self.mpu_partcopy_bkt,
-                                                                         self.mpu_partcopy_obj,
-                                                                         parts=uploaded_parts2)
-        assert_utils.assert_true(status, f"Failed to upload parts: {new_parts}")
-        response = self.s3_mpu_test_obj.upload_part_copy(self.bucket_name/self.object_name,
-                                                             self.mpu_partcopy_bkt,
-                                                             self.mpu_partcopy_obj,
-                                                             part_number=2,
-                                                             upload_id=mpu_id2)
-        new_parts.append({"PartNumber": 2, "ETag": response["ETag"]})
-        self.log.info("Listing parts of multipart upload")
-        res = self.s3_mpu_test_obj.list_parts(mpu_id2, self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        assert_utils.assert_true(res[0], res[1])
-        self.log.info("Listed parts of multipart upload: %s", res[1])
-        self.log.info("Complete the multipart")
-        try:
-            resp = self.s3_mpu_test_obj.complete_multipart_upload(mpu_id2, new_parts[1],
-                                                                  self.mpu_partcopy_bkt,
-                                                                  self.mpu_partcopy_obj)
-            assert_utils.assert_false(resp[0], resp[1])
-        except CTException as error:
-            self.log.error(error)
-            self.log.info("Failed to complete the multipart with provided part details ")
-        upload_etag2 = resp[1]["ETag"]
-        self.log.info("Get the uploaded object (using uploadPartCopy)")
-        status, res = self.s3_test_obj.get_object(self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        assert_utils.assert_true(status, res)
-        get_etag2 = res['ETag']
-        self.log.info("Compare ETags")
-        assert_utils.assert_equal(resp[1]["ETag"], get_etag2,
-                                  f"Failed to match ETag: {upload_etag2}, {get_etag2}")
-        self.log.info("Matched ETag: %s, %s", upload_etag2, get_etag2)
-
-        self.log.info("Stop and validate parallel S3 IOs")
-        s3_background_io.stop()
-        s3_background_io.cleanup()
-        self.log.info("ENDED: Get the object uploaded using MPU UploadPartCopy")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32711')
-    @CTFailOn(error_handler)
-    def test_mpu_upload_partcopy_32711(self):
-        """
-        This test is for copying uploaded using multipart UploadPartCopy api to another buckets
-        """
-        mp_config = MPART_CFG["test_32702_1"]
-        multipart_obj_path = self.mp_obj_path
-        if os.path.exists(multipart_obj_path):
-            os.remove(multipart_obj_path)
-        create_file(multipart_obj_path, mp_config["file_size"])
-
-        self.log.info("STARTED: Test Copy object created by MPU uploadPartCopy to another bucket")
-        s3_background_io = S3BackgroundIO(s3_test_lib_obj=self.s3_test_obj)
-        self.log.info("start s3 IO's")
-        s3_background_io.start(log_prefix="TEST-32711_s3bench_ios", duration="0h2m")
-        # create and complete MPU for Bucket1/object1
-        self.create_and_complete_mpu(mp_config)
-        # upload anotherobject using simple put operation
-        self.put_object_name = "mpu-uploadpartcopy-put-obj"
-        status, put_etag = self.s3_test_obj.put_object(self.bucket_name, self.put_object_name,
-                                                       multipart_obj_path)
-        assert_utils.assert_true(status, put_etag)
-        # create and complete MPU for Bucket2/object2
-        self.log.info("Creating a bucket with name : %s", self.mpu_partcopy_bkt)
-        res = self.s3_test_obj.create_bucket(self.mpu_partcopy_bkt)
-        assert_utils.assert_true(res[0], res[1])
-        assert_utils.assert_equal(res[1], self.mpu_partcopy_bkt, res[1])
-        self.log.info("Created a bucket with name : %s", self.mpu_partcopy_bkt)
-        mp_config_2 = MPART_CFG["test_32702_2"]
-        self.create_file_mpu(mp_config_2["file_size"], self.mp_obj_path_partcopy)
-        mpu_id2 = self.initiate_multipart(self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        uploaded_parts2 = get_precalculated_parts(self.mp_obj_path_partcopy,
-                                                  mp_config_2["part_sizes"],
-                                                  chunk_size=mp_config_2["chunk_size"])
-        keys = list(uploaded_parts2.keys())
-        uploaded_parts2['3'] = uploaded_parts2.pop('2')  # modified the part 2 as part 3
-        random.shuffle(keys)
-        self.log.info("Uploading parts")
-        status, new_parts = self.s3_mpu_test_obj.upload_parts_sequential(mpu_id2,
-                                                                         self.mpu_partcopy_bkt,
-                                                                         self.mpu_partcopy_obj,
-                                                                         parts=uploaded_parts2)
-        assert_utils.assert_true(status, f"Failed to upload parts: {new_parts}")
-        response = self.s3_mpu_test_obj.upload_part_copy(self.bucket_name / self.object_name,
-                                                         self.mpu_partcopy_bkt,
-                                                         self.mpu_partcopy_obj,
-                                                         part_number=2,
-                                                         upload_id=mpu_id2)
-        new_parts.append({"PartNumber": 2, "ETag": response["ETag"]})
-        response1 = self.s3_mpu_test_obj.upload_part_copy(self.bucket_name / self.put_object_name,
-                                                         self.mpu_partcopy_bkt,
-                                                         self.mpu_partcopy_obj,
-                                                         part_number=4,
-                                                         upload_id=mpu_id2)
-        new_parts.append({"PartNumber": 4, "ETag": response["ETag"]})
-        self.log.info("Listing parts of multipart upload")
-        res = self.s3_mpu_test_obj.list_parts(mpu_id2, self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        assert_utils.assert_true(res[0], res[1])
-        self.log.info("Listed parts of multipart upload: %s", res[1])
-        self.log.info("Complete the multipart")
-        try:
-            resp = self.s3_mpu_test_obj.complete_multipart_upload(mpu_id2, new_parts[1],
-                                                                  self.mpu_partcopy_bkt,
-                                                                  self.mpu_partcopy_obj)
-            assert_utils.assert_false(resp[0], resp[1])
-        except CTException as error:
-            self.log.error(error)
-            self.log.info("Failed to complete the multipart with provided part details ")
-        upload_etag2 = resp[1]["ETag"]
-        self.log.info("Get the uploaded object (using uploadPartCopy)")
-        status, res = self.s3_test_obj.get_object(self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        assert_utils.assert_true(status, res)
-        get_etag2 = res['ETag']
-        self.log.info("Compare ETags")
-        assert_utils.assert_equal(resp[1]["ETag"], get_etag2,
-                                  f"Failed to match ETag: {upload_etag2}, {get_etag2}")
-        self.log.info("Matched ETag: %s, %s", upload_etag2, get_etag2)
-        # copy object2 to bucket3 and bucket1
-        src_bkt = self.mpu_partcopy_bkt
-        dst_bkt = "mp-bkt3-{}".format(perf_counter_ns())
-        self.log.info("Creating a bucket with name : %s", dst_bkt)
-        resp = self.s3_test_obj.create_bucket(dst_bkt)
-        assert_utils.assert_true(resp[0], resp[1])
-        assert_utils.assert_equal(resp[1], dst_bkt, resp[1])
-        self.log.info("Created a bucket with name : %s", dst_bkt)
-        self.log.info("Copy object from bucket2 named %s to bucket3 named %s", src_bkt, dst_bkt)
-        resp = self.s3_test_obj.copy_object(
-            source_bucket=src_bkt, source_object=self.mpu_partcopy_obj, dest_bucket=dst_bkt,
-            dest_object=self.mpu_partcopy_obj)
-        assert_utils.assert_true(resp[0], resp[1])
-        copy_etag = resp[1]['CopyObjectResult']['ETag']
-        self.log.info("Verify copy and source etags match")
-        assert_utils.assert_equal(upload_etag2, copy_etag)
-        dst_bkt = self.bucket_name
-        self.log.info("Copy object from bucket2 named %s to bucket1 named %s", src_bkt, dst_bkt)
-        resp = self.s3_test_obj.copy_object(
-            source_bucket=src_bkt, source_object=self.mpu_partcopy_obj, dest_bucket=dst_bkt,
-            dest_object=self.mpu_partcopy_obj)
-        assert_utils.assert_true(resp[0], resp[1])
-        copy_etag = resp[1]['CopyObjectResult']['ETag']
-        self.log.info("Verify copy and source etags match")
-        assert_utils.assert_equal(upload_etag2, copy_etag)
-        self.log.info("Stop and validate parallel S3 IOs")
-        s3_background_io.stop()
-        s3_background_io.cleanup()
-        self.log.info("ENDED: Test Copy object created by MPU uploadPartCopy to another bucket")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32710')
-    @CTFailOn(error_handler)
-    def test_mpu_upload_partcopy_32710(self):
-        """
-        This test is for Copy object created recursively by MPU-UploadPartCopy
-        """
-        mp_config = MPART_CFG["test_32702_1"]
-        multipart_obj_path = self.mp_obj_path
-        if os.path.exists(multipart_obj_path):
-            os.remove(multipart_obj_path)
-        create_file(multipart_obj_path, mp_config["file_size"])
-        mp_config_2 = MPART_CFG["test_32702_2"]
-        if os.path.exists(self.mp_obj_path_partcopy):
-            os.remove(self.mp_obj_path_partcopy)
-        create_file(self.mp_obj_path_partcopy, mp_config_2["file_size"])
-        self.log.info("STARTED: Test Multipart upload with invalid json input")
-        s3_background_io = S3BackgroundIO(s3_test_lib_obj=self.s3_test_obj)
-        self.log.info("start s3 IO's")
-        s3_background_io.start(log_prefix="TEST-32710_s3bench_ios", duration="0h2m")
-        # create and complete MPU for Bucket1/object1
-        self.create_and_complete_mpu(mp_config)
-        self.log.info("Creating a bucket with name : %s", self.mpu_partcopy_bkt)
-        res = self.s3_test_obj.create_bucket(self.mpu_partcopy_bkt)
-        assert_utils.assert_true(res[0], res[1])
-        assert_utils.assert_equal(res[1], self.mpu_partcopy_bkt, res[1])
-        self.log.info("Created a bucket with name : %s", self.mpu_partcopy_bkt)
-        mpu_id2 = self.initiate_multipart(self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        uploaded_parts2 = get_precalculated_parts(self.mp_obj_path_partcopy,
-                                                  mp_config_2["part_sizes"],
-                                                  chunk_size=mp_config_2["chunk_size"])
-        keys = list(uploaded_parts2.keys())
-        uploaded_parts2.pop('2')  #removed part 2 as we are going to uplaod only one part here
-        self.log.info("Uploading parts")
-        status, new_parts = self.s3_mpu_test_obj.upload_parts_sequential(mpu_id2,
-                                                                         self.mpu_partcopy_bkt,
-                                                                         self.mpu_partcopy_obj,
-                                                                         parts=uploaded_parts2)
-        assert_utils.assert_true(status, f"Failed to upload parts: {new_parts}")
-        response = self.s3_mpu_test_obj.upload_part_copy(self.bucket_name / self.object_name,
-                                                         self.mpu_partcopy_bkt,
-                                                         self.mpu_partcopy_obj,
-                                                         part_number=2,
-                                                         upload_id=mpu_id2)
-        new_parts.append({"PartNumber": 2, "ETag": response["ETag"]})
-        self.log.info("Listing parts of multipart upload")
-        res = self.s3_mpu_test_obj.list_parts(mpu_id2, self.mpu_partcopy_bkt, self.mpu_partcopy_obj)
-        assert_utils.assert_true(res[0], res[1])
-        self.log.info("Listed parts of multipart upload: %s", res[1])
-        self.log.info("Complete the multipart")
-        try:
-            resp = self.s3_mpu_test_obj.complete_multipart_upload(mpu_id2, new_parts[1],
-                                                                  self.mpu_partcopy_bkt,
-                                                                  self.mpu_partcopy_obj)
-            assert_utils.assert_false(resp[0], resp[1])
-        except CTException as error:
-            self.log.error(error)
-            self.log.info("Failed to complete the multipart with provided part details ")
-        self.log.info("Creating a bucket with name : %s", self.mpu_partcopy_bkt)
-        res = self.s3_test_obj.create_bucket(self.mpu_partcopy_bkt)
-        assert_utils.assert_true(res[0], res[1])
-        assert_utils.assert_equal(res[1], self.mpu_partcopy_bkt, res[1])
-        self.log.info("Created a bucket with name : %s", self.mpu_partcopy_bkt)
-        mpu_partcopy_bkt3 = "mpu-partcopy-bkt3".format(perf_counter_ns())
-        mpu_partcopy_obj3 = "mpu-partcopy-obj3".format(perf_counter_ns())
-        self.log.info("Creating a bucket with name : %s", mpu_partcopy_bkt3)
-        res = self.s3_test_obj.create_bucket(mpu_partcopy_bkt3)
-        assert_utils.assert_true(res[0], res[1])
-        assert_utils.assert_equal(res[1], mpu_partcopy_bkt3, res[1])
-        self.log.info("Created a bucket with name : %s", mpu_partcopy_bkt3)
-        mpu_id3 = self.initiate_multipart(mpu_partcopy_bkt3, mpu_partcopy_obj3)
-        uploaded_parts2 = get_precalculated_parts(self.mp_obj_path_partcopy,
-                                                  mp_config_2["part_sizes"],
-                                                  chunk_size=mp_config_2["chunk_size"])
-        keys = list(uploaded_parts2.keys())
-        uploaded_parts2.pop('2')  # removed part 2 as we are going to uplaod only one part here
-        self.log.info("Uploading parts")
-        status, new_parts = self.s3_mpu_test_obj.upload_parts_sequential(mpu_id3,
-                                                                         mpu_partcopy_bkt3,
-                                                                         mpu_partcopy_obj3,
-                                                                         parts=uploaded_parts2)
-        assert_utils.assert_true(status, f"Failed to upload parts: {new_parts}")
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            self.mpu_partcopy_bkt / self.mpu_partcopy_obj, mpu_partcopy_bkt3,
-            mpu_partcopy_obj3, part_number=2, upload_id=mpu_id3)
-        new_parts.append({"PartNumber": 2, "ETag": response["ETag"]})
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            self.bucket_name / self.object_name, mpu_partcopy_bkt3,
-            mpu_partcopy_obj3, part_number=3, upload_id=mpu_id3)
-        new_parts.append({"PartNumber": 3, "ETag": response["ETag"]})
-        self.log.info("Listing parts of multipart upload")
-        res = self.s3_mpu_test_obj.list_parts(mpu_id3, mpu_partcopy_bkt3, mpu_partcopy_obj3)
-        assert_utils.assert_true(res[0], res[1])
-        self.log.info("Listed parts of multipart upload: %s", res[1])
-        self.log.info("Complete the multipart")
-        try:
-            resp = self.s3_mpu_test_obj.complete_multipart_upload(mpu_id3, new_parts[1],
-                                                                  mpu_partcopy_bkt3,
-                                                                  mpu_partcopy_obj3)
-            assert_utils.assert_false(resp[0], resp[1])
-        except CTException as error:
-            self.log.error(error)
-            self.log.info("Failed to complete the multipart with provided part details ")
-        self.log.info("Stop and validate parallel S3 IOs")
-        s3_background_io.stop()
-        s3_background_io.cleanup()
-        self.log.info("ENDED: Test Copy object created  recursively by MPU-UploadPartCopy")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32721')
-    @CTFailOn(error_handler)
-    def test_32721(self):
-        """Upload 2 objects parallelly using uploadPartCopy."""
-        self.log.info("STARTED: Upload 2 objects parallelly using uploadPartCopy.")
-        self.log.info("Start S3 background IOs.")
-        self.s3bio_obj.start(log_prefix="TEST-32721_s3bench_ios", duration="0h2m")
-        self.log.info("Step 1: Create bucket1.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name1)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 2: Initiate multipart upload by performing CreateMultipartUpload1.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id1 = resp[1]["UploadId"]
-        self.log.info("Step 3: Upload 10 parts of various sizes and in random order. The uploaded "
-                      "parts can have sizes aligned and unaligned with motr unit sizes. "
-                      "Make sure the entire object is > 5GB.")
-        resp = create_file(self.test_file, count=6, b_size="1G")
-        assert_utils.assert_true(resp[0], resp[1])
-        chunks = s3_utils.get_unaligned_parts(
-            self.test_file, total_parts=10, random=True)
-        status, parts = self.s3mpu_obj.upload_parts_sequential(
-            mpu_id1, self.bucket_name, self.object_name, parts=chunks)
-        assert_utils.assert_true(status, f"Failed to upload parts: {parts}")
-        self.log.info("Step 4: Do ListParts to see the parts uploaded.")
-        resp = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 5: Get the part details and perform CompleteMultipartUpload.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name, self.object_name)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 6: Create bucket2.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name2)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 7: Initiate multipart upload by performing CreateMultipartUpload2.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id2 = resp[1]["UploadId"]
-        self.log.info("Step 8: Initiate multipart upload by performing CreateMultipartUpload3.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id1 = resp[1]["UploadId"]
-        self.log.info("Step 9: Parallelly upload part1 by copying object1 to both these uploadIDs.")
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-            part_number=1, upload_id=mpu_id2, opy_source_range="'bytes=1-100000'")
-        assert_utils.assert_true(response[0], response)
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-            part_number=1, upload_id=mpu_id2, copy_source_range="'bytes=1-100000'")
-        assert_utils.assert_true(response[0], response)
-        self.log.info("Step 10: Perform ListParts prallely on these 2 uploadIDs.")
-        res1 = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(res1[0], res1[1])
-        res2 = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(res2[0], res2[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 11: Complete MPU for both uploadIDs.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Stop & validate S3 background IOs.")
-        self.s3bio_obj.stop()
-        self.s3bio_obj.validate()
-        self.log.info("ENDED: Upload 2 objects parallelly using uploadPartCopy.")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32730')
-    @CTFailOn(error_handler)
-    def test_32730(self):
-        """
-        Upload part copy.
-
-        Upload object UploadPartCopy having part sizes within aws part size limits using byte-range.
-        """
-        self.log.info("STARTED: Upload object UploadPartCopy using byte-range param")
-        self.log.info("Start S3 background IOs.")
-        self.s3bio_obj.start(log_prefix="TEST-32730_s3bench_ios", duration="0h2m")
-        self.log.info("Step 1: Create bucket1.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name1)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 2: Initiate multipart upload by performing CreateMultipartUpload.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id1 = resp[1]["UploadId"]
-        self.log.info("Step 3: Upload 10 parts of various sizes and in random order. The uploaded "
-                      "parts can have sizes aligned and unaligned with motr unit sizes."
-                      " Make sure the entire object is <= 5GB")
-        resp = create_file(self.test_file, count=9, b_size="533M")
-        assert_utils.assert_true(resp[0], resp[1])
-        chunks = s3_utils.get_unaligned_parts(
-            self.test_file, total_parts=10, random=True)
-        status, parts = self.s3mpu_obj.upload_parts_sequential(
-            mpu_id1, self.bucket_name, self.object_name, parts=chunks)
-        assert_utils.assert_true(status, f"Failed to upload parts: {parts}")
-        self.log.info("Step 4: Do ListParts to see the parts uploaded.")
-        resp = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 5: Get the part details and perform CompleteMultipartUpload.")
-        parts = resp[1]["Parts"]
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name, self.object_name)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 6: Create bucket2.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name2)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 7: Initiate multipart upload by performing CreateMultipartUpload2.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id2 = resp[1]["UploadId"]
-        self.log.info("Step 8: Upload part1 by copying object1 to it.")
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-            part_number=1, upload_id=mpu_id2)
-        assert_utils.assert_true(response[0], response)
-        self.log.info("Step 9: Upload part part 2, 3 using simple upload.")
-        resp = self.s3mpu_obj.upload_part(
-            os.urandom(5000), self.bucket_name, self.object_name, upload_id=mpu_id2, part_number=2)
-        assert_utils.assert_true(resp[0], f"Failed to upload part2: {resp}")
-        resp = self.s3mpu_obj.upload_part(
-            os.urandom(3000), self.bucket_name, self.object_name, upload_id=mpu_id2, part_number=3)
-        assert_utils.assert_true(resp[0], f"Failed to upload part3: {resp}")
-        self.log.info("Step 10: list the parts uploaded.")
-        resp = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 11: Complete the MPU.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Stop & validate S3 background IOs.")
-        self.s3bio_obj.stop()
-        self.s3bio_obj.validate()
-        self.log.info("ENDED: Upload object UploadPartCopy using byte-range param")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32734')
-    @CTFailOn(error_handler)
-    def test_32734(self):
-        """Overwrite the failed UploadCopyPart for byte range to successful uploadPartCopy."""
-        self.log.info("STARTED: Overwrite the failed UploadCopyPart for byte range to successful.")
-        self.log.info("Start S3 background IOs.")
-        self.s3bio_obj.start(log_prefix="TEST-32734_s3bench_ios", duration="0h2m")
-        self.log.info("Step 1: Create bucket1.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name1)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 2: Initiate multipart upload by performing CreateMultipartUpload.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id1 = resp[1]["UploadId"]
-        self.log.info("Step 3: Upload 10 parts of various sizes and in random order. The uploaded "
-                      "parts can have sizes aligned and unaligned with motr unit sizes."
-                      " Make sure the entire object is > 5GB.")
-        resp = create_file(self.test_file, count=11, b_size="1G")
-        assert_utils.assert_true(resp[0], resp[1])
-        chunks = s3_utils.get_unaligned_parts(self.test_file, total_parts=10, random=True)
-        status, parts = self.s3mpu_obj.upload_parts_sequential(
-            mpu_id1, self.bucket_name, self.object_name, parts=chunks)
-        assert_utils.assert_true(status, f"Failed to upload parts: {parts}")
-        self.log.info("Step 4: Do ListParts to see the parts uploaded.")
-        resp = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 5: Get the part details and perform CompleteMultipartUpload.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name, self.object_name)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 6: Create bucket2.")
-        resp = self.s3t_obj.create_bucket(self.bucket_name1)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 7: Initiate multipart upload by performing CreateMultipartUpload.")
-        resp = self.s3mpu_obj.create_multipart_upload(self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id2 = resp[1]["UploadId"]
-        self.log.info("Step 8: Upload part2 by copying entire object1 >5GB to uploadID2.")
-        try:
-            response = self.s3_mpu_test_obj.upload_part_copy(
-                f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-                part_number=2, upload_id=mpu_id2)
-            assert_utils.assert_false(response[0], response)
-        except CTException as error:
-            assert_utils.assert_in("InvalidParts", error.message)
-        self.log.info("Step 9: Upload part1 by copying byte range > 5MB from object1 to uploadID2.")
-        try:
-            response = self.s3_mpu_test_obj.upload_part_copy(
-                f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-                part_number=1, upload_id=mpu_id2, copy_source_range="'bytes=1-10'")
-            assert_utils.assert_false(response[0], response)
-        except CTException as error:
-            assert_utils.assert_in("InvalidParts", error.message)
-        self.log.info("Step 10: Upload part2 by normal file upload.")
-        resp = self.s3mpu_obj.upload_part(os.urandom(1000), self.bucket_name2,
-                                          self.object_name2, part_number=2, upload_id=mpu_id2)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 11: Upload part1 by copying byte range within upload part size limits.")
-        response = self.s3_mpu_test_obj.upload_part_copy(
-            f"{self.bucket_name}/{self.object_name}", self.bucket_name2, self.object_name2,
-            part_number=1, upload_id=mpu_id2, copy_source_range="'bytes=1-1073741824'")
-        assert_utils.assert_false(response[0], response)
-        self.log.info("Step 12: list the parts uploaded.")
-        resp = self.s3mpu_obj.list_parts(mpu_id1, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 13: Complete the MPU.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = self.s3mpu_obj.complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Stop & validate S3 background IOs.")
-        self.s3bio_obj.stop()
-        self.s3bio_obj.validate()
-        self.log.info("ENDED: Overwrite the failed UploadCopyPart for byte range to successful.")
-
-    @pytest.mark.s3_ops
-    @pytest.mark.tags('TEST-32737')
-    @CTFailOn(error_handler)
-    def test_32737(self):
-        """Uploading part using UploadPartCopy which is from another account."""
-        self.log.info("STARTED: Uploading part using UploadPartCopy which is from another account.")
-        self.log.info("Start S3 background IOs.")
-        self.s3bio_obj.start(log_prefix="TEST-32737_s3bench_ios", duration="0h2m")
-        s3_account1 = self.acc_name_prefix.format(perf_counter_ns())
-        response = cmn_lib.create_s3_account_get_s3lib_objects(
-            s3_account1, self.acc_email_prefix.format(s3_account1), self.s3acc_passwd)
-        self.account_list.append(s3_account1)
-        self.log.info("Step 1: Create bucket1")
-        resp = response[1].create_bucket(self.bucket_name1)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 2: Initiate multipart upload by performing CreateMultipartUpload.")
-        resp = response[8].create_multipart_upload(self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id1 = resp[1]["UploadId"]
-        self.log.info("Step 3: Upload 10 parts of various sizes and in random order.")
-        resp = create_file(self.test_file, count=5, b_size="1G")
-        assert_utils.assert_true(resp[0], resp[1])
-        chunks = s3_utils.get_unaligned_parts(
-            self.test_file, total_parts=10, random=True)
-        status, parts = response[8].upload_parts_sequential(
-            mpu_id1, self.bucket_name, self.object_name, parts=chunks)
-        assert_utils.assert_true(status, f"Failed to upload parts: {parts}")
-        self.log.info("Step 4: Do ListParts to see the parts uploaded.")
-        resp = response[8].list_parts(mpu_id1, self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 5: Get the part details and perform CompleteMultipartUpload.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = response[8].complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name, self.object_name)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 6: Create bucket2 in account 2 and have the default settings. Meaning,"
-                      " account account 2 does not have read permission on account1’s resources.")
-        s3_account2 = self.acc_name_prefix.format(perf_counter_ns())
-        response1 = cmn_lib.create_s3_account_get_s3lib_objects(
-            s3_account2, self.acc_email_prefix.format(s3_account2), self.s3acc_passwd)
-        self.account_list.append(s3_account2)
-        self.log.info("Step 7: Initiate multipart upload by performing CreateMultipartUpload.")
-        resp = response[8].create_multipart_upload(self.bucket_name1, self.object_name1)
-        assert_utils.assert_true(resp[0], resp[1])
-        mpu_id2 = resp[1]["UploadId"]
-        self.log.info("Step 8: Upload part 2 by copying entire object 1 to uploadID 2.")
-        try:
-            response = response1[8].upload_part_copy(
-                f"{self.bucket_name1}/{self.object_name1}", self.bucket_name2, self.object_name2,
-                part_number=1, upload_id=mpu_id2, copy_source_range="'bytes=1-10'")
-            assert_utils.assert_false(response[0], response)
-        except CTException as error:
-            assert_utils.assert_in("InvalidParts", error.message)
-        self.log.info("Step 9: Upload part 1 by using regular file upload.")
-        resp = response1[8].upload_part(os.urandom(1000), self.bucket_name2,
-                                        self.object_name2, part_number=2, upload_id=mpu_id2)
-        assert_utils.assert_true(resp[0], resp)
-        self.log.info("Step 10: Give read permission to account 2 on account 1.")
-        resp = response1[2].put_bucket_acl(self.bucket_name1,
-                                           grant_read="{}{}".format("id=", response1[0]))
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Step 11: After that upload part 2 by copying entire object 1.")
-        response = response1[8].upload_part_copy(
-            f"{self.bucket_name1}/{self.object_name1}", self.bucket_name2, self.object_name2,
-            part_number=2, upload_id=mpu_id2)
-        assert_utils.assert_true(response[0], response)
-        self.log.info("Step 12: list the parts uploaded.")
-        resp = response1[8].list_parts(mpu_id1, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        parts = resp[1]["Parts"]
-        self.log.info("Parts: %s", parts)
-        self.log.info("Step 13: Complete the MPU.")
-        sorted_part_list = sorted(parts, key=lambda x: x['PartNumber'])
-        resp = response1[8].complete_multipart_upload(
-            mpu_id1, sorted_part_list, self.bucket_name2, self.object_name2)
-        assert_utils.assert_true(resp[0], resp[1])
-        self.log.info("Stop & validate S3 background IOs.")
-        self.s3bio_obj.stop()
-        self.s3bio_obj.validate()
-        self.log.info("ENDED: Uploading part using UploadPartCopy which is from another account.")
